refactor infix postfix
