+++ types/infix_postfix_test.go
@@ -1,4 +1,4 @@
-package mpc
+package types
 
 import (
 	"fmt"
